NLP/task_1/requirements.txt
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
dialogflow==0.7.2
future==0.18.2
google-api-core==1.14.3
google-auth==1.7.1
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
protobuf==3.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
tornado==6.0.3
urllib3==1.25.7
